requirements.txt
Flask
Flask-Cors
networkx
numpy
requests
pandas
scikit-learn
scipy
python-chess
matplotlib
tqdm